requirements.txt
flask
flask_cors
pyyaml
numpy
gensim
sentence_transformers